LayoutTests/imported/w3c/web-platform-tests/html/infrastructure/safe-passing-of-structured-data/shared-array-buffers/window-failure.https-expected.txt

CONSOLE MESSAGE: line 2: ReferenceError: Can't find variable: SharedArrayBuffer

Harness Error (TIMEOUT), message = null

TIMEOUT SharedArrayBuffer and a cross-site <iframe> Test timed out
